Fix exception thrown when run from command line
Opening project from command line doesn't work because exception is
thrown when trying to launch the application. Exception occurs because
application is trying to retrieve _launchArguments for the row of the
currently selected cell. When running from command line currently
selected cell is null and that causes a null reference exception to
occur.

Issue is fixed by checking the current grid and cell for null and
handling the situation like no row has been selected.